# Update go-bitfield