--- a/shared/aggregation/aggregation.go
+++ b/shared/aggregation/aggregation.go
@@ -14,9 +14,6 @@ var (
 	// ErrBitsOverlap is returned when two bitlists overlap with each other.
 	ErrBitsOverlap = errors.New("overlapping aggregation bits")
 
-	// ErrBitsDifferentLen is returned when two bitlists have different lengths.
-	ErrBitsDifferentLen = errors.New("different bitlist lengths")
-
 	// ErrInvalidStrategy is returned when invalid aggregation strategy is selected.
 	ErrInvalidStrategy = errors.New("invalid aggregation strategy")
 )