--- a/shared/aggregation/attestations/attestations.go
+++ b/shared/aggregation/attestations/attestations.go
@@ -66,9 +66,6 @@ func Aggregate(atts []*ethpb.Attestation) ([]*ethpb.Attestation, error) {
 
 // AggregatePair aggregates pair of attestations a1 and a2 together.
 func AggregatePair(a1, a2 *ethpb.Attestation) (*ethpb.Attestation, error) {
-	if a1.AggregationBits.Len() != a2.AggregationBits.Len() {
-		return nil, aggregation.ErrBitsDifferentLen
-	}
 	o, err := a1.AggregationBits.Overlaps(a2.AggregationBits)
 	if err != nil {
 		return nil, err