… beacon-chain/operations/attestations/kv/block_test.go
… beacon-chain/rpc/prysm/v1alpha1/validator/attester_test.go
… beacon-chain/rpc/prysm/v1alpha1/validator/proposer_attestations_test.go
… beacon-chain/rpc/prysm/v1alpha1/validator/proposer_sync_aggregate_test.go
… beacon-chain/rpc/prysm/v1alpha1/validator/proposer_test.go
… beacon-chain/rpc/prysm/v1alpha1/validator/server_test.go
… go.sum
… shared/aggregation/attestations/attestations_bench_test.go
… shared/aggregation/attestations/attestations_test.go
… shared/aggregation/attestations/maxcover_test.go
… shared/aggregation/maxcover.go
… shared/aggregation/maxcover_test.go