**What type of PR is this?**

Other

**What does this PR do? Why is it needed?**

Updates go-bitfield dependency to the latest changes to remove all panic paths.

**Which issues(s) does this PR fix?**

**Other notes for review**

This PR changes a little bit of the logic within the max cover attestation aggregation where a list of attestations with different bitfield lengths would fail silently and simply return the original list. I think we should review the usage of these routines to ensure that only validated inputs are possible.


## Benchmarks (no expected improvement or regression)

**BEFORE**
```
goos: linux
goarch: amd64
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
BenchmarkAggregateAttestations_Aggregate
    bitlistutils.go:49: bitlistsWithMultipleBitSet random seed: 1625756773692161169
BenchmarkAggregateAttestations_Aggregate/naive_256_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/naive_256_attestations_with_single_bit_set-8                  9         124738407 ns/op          800268 B/op       5870 allocs/op
BenchmarkAggregateAttestations_Aggregate/max-cover_256_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/max-cover_256_attestations_with_single_bit_set-8             14          82141123 ns/op        10153576 B/op      37002 allocs/op
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_256_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_256_attestations_with_single_bit_set-8         19          63808336 ns/op          435481 B/op       2838 allocs/op
BenchmarkAggregateAttestations_Aggregate/naive_256_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/naive_256_attestations_with_64_random_bits_set-8              9         130368631 ns/op          800268 B/op       5870 allocs/op
BenchmarkAggregateAttestations_Aggregate/max-cover_256_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/max-cover_256_attestations_with_64_random_bits_set-8         13          90312877 ns/op        10153778 B/op      37002 allocs/op
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_256_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_256_attestations_with_64_random_bits_set-8                     15          70614499 ns/op          435483 B/op       2838 allocs/op
BenchmarkAggregateAttestations_Aggregate/naive_512_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/naive_512_attestations_with_single_bit_set-8                                  4         255543859 ns/op         1603088 B/op      11758 allocs/op
BenchmarkAggregateAttestations_Aggregate/max-cover_512_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/max-cover_512_attestations_with_single_bit_set-8                              5         215740544 ns/op        39180929 B/op     139530 allocs/op
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_512_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_512_attestations_with_single_bit_set-8                          8         148293825 ns/op          868730 B/op       5654 allocs/op
BenchmarkAggregateAttestations_Aggregate/naive_1024_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/naive_1024_attestations_with_64_random_bits_set-8                             3         384351466 ns/op         2422704 B/op      17682 allocs/op
BenchmarkAggregateAttestations_Aggregate/max-cover_1024_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/max-cover_1024_attestations_with_64_random_bits_set-8                         2         668229163 ns/op        153955828 B/op    663455 allocs/op
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_1024_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_1024_attestations_with_64_random_bits_set-8                     4         294898976 ns/op         3626320 B/op      16932 allocs/op
PASS

```

**AFTER**
```
goos: linux
goarch: amd64
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
BenchmarkAggregateAttestations_Aggregate
    bitlistutils.go:49: bitlistsWithMultipleBitSet random seed: 1625756655775133188
BenchmarkAggregateAttestations_Aggregate/naive_256_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/naive_256_attestations_with_single_bit_set-8                  9         125408339 ns/op          800279 B/op       5870 allocs/op
BenchmarkAggregateAttestations_Aggregate/max-cover_256_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/max-cover_256_attestations_with_single_bit_set-8             14          81169746 ns/op        10153733 B/op      37002 allocs/op
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_256_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_256_attestations_with_single_bit_set-8         19          60975770 ns/op          435481 B/op       2838 allocs/op
BenchmarkAggregateAttestations_Aggregate/naive_256_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/naive_256_attestations_with_64_random_bits_set-8              9         127435233 ns/op          800268 B/op       5870 allocs/op
BenchmarkAggregateAttestations_Aggregate/max-cover_256_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/max-cover_256_attestations_with_64_random_bits_set-8         14          81103776 ns/op        10153577 B/op      37002 allocs/op
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_256_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_256_attestations_with_64_random_bits_set-8                     18          66603796 ns/op          435481 B/op       2838 allocs/op
BenchmarkAggregateAttestations_Aggregate/naive_512_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/naive_512_attestations_with_single_bit_set-8                                  5         242065371 ns/op         1603088 B/op      11758 allocs/op
BenchmarkAggregateAttestations_Aggregate/max-cover_512_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/max-cover_512_attestations_with_single_bit_set-8                              5         209957859 ns/op        39180931 B/op     139530 allocs/op
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_512_attestations_with_single_bit_set
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_512_attestations_with_single_bit_set-8                          8         129007576 ns/op          868730 B/op       5654 allocs/op
BenchmarkAggregateAttestations_Aggregate/naive_1024_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/naive_1024_attestations_with_64_random_bits_set-8                             3         356786276 ns/op         2427808 B/op      17720 allocs/op
BenchmarkAggregateAttestations_Aggregate/max-cover_1024_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/max-cover_1024_attestations_with_64_random_bits_set-8                         3         474591620 ns/op        153064802 B/op    660272 allocs/op
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_1024_attestations_with_64_random_bits_set
BenchmarkAggregateAttestations_Aggregate/opt-max-cover_1024_attestations_with_64_random_bits_set-8                     4         299720032 ns/op         3574532 B/op      16841 allocs/op
PASS
```